More cleanup of the schema crate
