+++ source/postcard-schema/src/impls/heapless_v0_8.rs
@@ -1,5 +1,7 @@
+//! Implementations of the [`Schema`] trait for the `heapless` crate v0.8
+
 use crate::{
-    schema::{NamedType, DataModelType},
+    schema::{DataModelType, NamedType},
     Schema,
 };
