+++ source/postcard-schema/src/impls/nalgebra_v0_33.rs
@@ -1,3 +1,5 @@
+//! Implementations of the [`Schema`] trait for the `nalgebra` crate v0.33
+
 use crate::{
     schema::{DataModelType, NamedType},
     Schema,
