+++ source/postcard-schema/src/impls/uuid_v1_0.rs
@@ -1,5 +1,7 @@
+//! Implementations of the [`Schema`] trait for the `uuid` crate v1.0
+
 use crate::{
-    schema::{NamedType, DataModelType},
+    schema::{DataModelType, NamedType},
     Schema,
 };
